requirements.txt
Django==4.2.4
django-environ==0.4.5
djangorestframework==3.14.0
psycopg2==2.9.1
requests == 2.31.0
dj-database-url==2.0.0
django-heroku==0.3.1
gunicorn==21.2.0
python-decouple==3.8
